+++ GuaranteedRate.Sextant/WebClients/SecureAsyncEventReporter.cs
@@ -1,22 +1,20 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Net;
+using System.Net.Http.Headers;
 
 namespace GuaranteedRate.Sextant.WebClients
 {
     public class SecureAsyncEventReporter : AsyncEventReporter
     {
-        public string authorization { get; set; }
-
-        protected override void ExtraSetup(WebRequest webRequest)
-        {
-            webRequest.Headers.Add("Authorization:" + authorization);
-        }
-
-        public SecureAsyncEventReporter(string url, int queueSize = DEFAULT_QUEUE_SIZE, int retries = DEFAULT_RETRIES)
+        public SecureAsyncEventReporter(string url, string authorization, int queueSize = DEFAULT_QUEUE_SIZE,
+            int retries = DEFAULT_RETRIES)
             : base(url, queueSize, retries)
         {
+            if (!string.IsNullOrEmpty(authorization))
+            {
+                Client.DefaultRequestHeaders.Authorization = AuthenticationHeaderValue.Parse(authorization);
+            }
         }
     }
 }